requirements.txt
hydra-core==1.3.2
scikit-learn==1.5.1
numpy==2.1.1
matplotlib==3.9.2
fpdf==1.7.2
